+++ CHANGES.md
@@ -115,6 +115,7 @@ Improvements:
 New language:
 
 - *MikroTik* RouterOS Scripting language by [Ivan Dementev][].
+- Google *Closure Templates* language by [Chema Balsas][]
 
 New style: